+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Trip api
 - Nearest api
 
+## [0.0.3] - 2021-02-09
+### Changed
+- The `route` function has been fully exposed so to allow also obtaining further
+  route annotations (e.g., speeds, distances, etc.)
+
 ## [0.0.2] - 2020-03-08
 ### Changed
 - Now path parameter is optional, and you can use either a valid path or
@@ -26,3 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Status enum
 - Working examples on readme and tests
 - test data unprocessed (raw) and preprocessed for CH and MLD
+
+## [0.1.0] - 2023-04-02
+### Added
+- Nearest and Table methods
+- Updated test files to osrm version 5.27.1